requirements.txt
about-time==3.1.1
aiogram==2.21
aiohttp==3.8.1
aiosignal==1.2.0
alive-progress==2.4.1
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
frozenlist==1.3.0
grapheme==0.6.0
h11==0.13.0
idna==3.3
loguru==0.6.0
lxml==4.9.1
multidict==6.0.2
outcome==1.2.0
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dotenv==0.20.0
pytube==12.1.0
pytz==2022.1
requests==2.28.1
selenium==4.3.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.21.0
trio-websocket==0.9.2
undetected-chromedriver==3.1.5.post4
urllib3==1.26.10
webdriver-manager==3.8.2
websockets==10.3
wsproto==1.1.0
yarl==1.7.2
